# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Samurai Resto API - Restaurant ordering system
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MenuCategory {
  id: string;
  name: string;
  /** @nullable */
  description?: string | null;
  sortOrder: number;
}

export interface MenuItem {
  id: string;
  sku: string;
  name: string;
  /** @nullable */
  description?: string | null;
  category: string;
  price: number;
  /** @nullable */
  imageUrl?: string | null;
  available: boolean;
  featured: boolean;
}

export interface OrderLineInput {
  menuItemId: string;
  quantity: number;
  /** @nullable */
  specialInstructions?: string | null;
}

export type OrderInputOrderType = typeof OrderInputOrderType[keyof typeof OrderInputOrderType];


export const OrderInputOrderType = {
  pickup: 'pickup',
  delivery: 'delivery',
} as const;

export interface OrderInput {
  customerName: string;
  customerPhone: string;
  /** @nullable */
  customerEmail?: string | null;
  orderType: OrderInputOrderType;
  /** @nullable */
  deliveryAddress?: string | null;
  items: OrderLineInput[];
  /** @nullable */
  specialInstructions?: string | null;
}

export interface OrderLine {
  menuItemId: string;
  menuItemName: string;
  quantity: number;
  unitPrice: number;
  subtotal: number;
  /** @nullable */
  specialInstructions?: string | null;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  preparing: 'preparing',
  ready: 'ready',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface Order {
  id: string;
  customerName: string;
  customerPhone: string;
  /** @nullable */
  customerEmail?: string | null;
  orderType: string;
  /** @nullable */
  deliveryAddress?: string | null;
  items: OrderLine[];
  subtotal: number;
  tax: number;
  total: number;
  status: OrderStatus;
  /** @nullable */
  squareOrderId?: string | null;
  /** @nullable */
  specialInstructions?: string | null;
  createdAt: string;
}

export interface ErrorResponse {
  error: string;
}

export type GetMenuItemsParams = {
category?: string;
};