# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Samurai Resto API - Restaurant ordering system
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: menu
    description: Menu items and categories
  - name: orders
    description: Online orders
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /menu/categories:
    get:
      operationId: getMenuCategories
      tags: [menu]
      summary: Get all menu categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuCategory"

  /menu/items:
    get:
      operationId: getMenuItems
      tags: [menu]
      summary: Get all menu items
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"

  /menu/featured:
    get:
      operationId: getFeaturedItems
      tags: [menu]
      summary: Get featured/popular menu items
      responses:
        "200":
          description: Featured items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"

  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    MenuCategory:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        sortOrder:
          type: integer
      required:
        - id
        - name
        - sortOrder

    MenuItem:
      type: object
      properties:
        id:
          type: string
        sku:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
        price:
          type: number
        imageUrl:
          type: ["string", "null"]
        available:
          type: boolean
        featured:
          type: boolean
      required:
        - id
        - sku
        - name
        - category
        - price
        - available
        - featured

    OrderLineInput:
      type: object
      properties:
        menuItemId:
          type: string
        quantity:
          type: integer
        specialInstructions:
          type: ["string", "null"]
      required:
        - menuItemId
        - quantity

    OrderInput:
      type: object
      properties:
        customerName:
          type: string
        customerPhone:
          type: string
        customerEmail:
          type: ["string", "null"]
        orderType:
          type: string
          enum: [pickup, delivery]
        deliveryAddress:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderLineInput"
        specialInstructions:
          type: ["string", "null"]
      required:
        - customerName
        - customerPhone
        - orderType
        - items

    OrderLine:
      type: object
      properties:
        menuItemId:
          type: string
        menuItemName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        subtotal:
          type: number
        specialInstructions:
          type: ["string", "null"]
      required:
        - menuItemId
        - menuItemName
        - quantity
        - unitPrice
        - subtotal

    Order:
      type: object
      properties:
        id:
          type: string
        customerName:
          type: string
        customerPhone:
          type: string
        customerEmail:
          type: ["string", "null"]
        orderType:
          type: string
        deliveryAddress:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderLine"
        subtotal:
          type: number
        tax:
          type: number
        total:
          type: number
        status:
          type: string
          enum: [pending, confirmed, preparing, ready, completed, cancelled]
        squareOrderId:
          type: ["string", "null"]
        specialInstructions:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - customerName
        - customerPhone
        - orderType
        - items
        - subtotal
        - tax
        - total
        - status
        - createdAt

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error