# lib/api-zod/src/generated/types/menuItem.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Samurai Resto API - Restaurant ordering system
 * OpenAPI spec version: 0.1.0
 */

export interface MenuItem {
  id: string;
  sku: string;
  name: string;
  /** @nullable */
  description?: string | null;
  category: string;
  price: number;
  /** @nullable */
  imageUrl?: string | null;
  available: boolean;
  featured: boolean;
}