# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Samurai Resto API - Restaurant ordering system
 * OpenAPI spec version: 0.1.0
 */
import type { OrderLine } from './orderLine';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: string;
  customerName: string;
  customerPhone: string;
  /** @nullable */
  customerEmail?: string | null;
  orderType: string;
  /** @nullable */
  deliveryAddress?: string | null;
  items: OrderLine[];
  subtotal: number;
  tax: number;
  total: number;
  status: OrderStatus;
  /** @nullable */
  squareOrderId?: string | null;
  /** @nullable */
  specialInstructions?: string | null;
  createdAt: string;
}