# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Samurai Resto API - Restaurant ordering system
 * OpenAPI spec version: 0.1.0
 */
import type { OrderInputOrderType } from './orderInputOrderType';
import type { OrderLineInput } from './orderLineInput';

export interface OrderInput {
  customerName: string;
  customerPhone: string;
  /** @nullable */
  customerEmail?: string | null;
  orderType: OrderInputOrderType;
  /** @nullable */
  deliveryAddress?: string | null;
  items: OrderLineInput[];
  /** @nullable */
  specialInstructions?: string | null;
}