# lib/api-zod/src/generated/types/orderLine.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Samurai Resto API - Restaurant ordering system
 * OpenAPI spec version: 0.1.0
 */

export interface OrderLine {
  menuItemId: string;
  menuItemName: string;
  quantity: number;
  unitPrice: number;
  subtotal: number;
  /** @nullable */
  specialInstructions?: string | null;
}